+++ docs/changelog.rst
@@ -11,6 +11,7 @@ Currently in development.
 - Added search to text fields with more than 10 choices (configurable with ``SAMPLEDB_MIN_NUM_TEXT_CHOICES_FOR_SEARCH``)
 - Added support for ``"first"`` and ``"last"`` statistic for timeseries
 - Allow using dicts for the ``style`` attribute
+- Added support for ``"include"`` style for object reference properties
 - Improved .eln export
 - Added support for file references as display properties
 - Allow configuring multiple workflow views